[CommonTools] move public header in to interface directory
