+++ CommonTools/RecoAlgos/plugins/ChargedRefCandidateProducer.cc
@@ -9,7 +9,7 @@
  *
  */
 
-#include "CommonTools/RecoAlgos/src/TrackToRefCandidate.h"
+#include "CommonTools/RecoAlgos/interface/TrackToRefCandidate.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
 
 typedef CandidateProducer<edm::View<reco::Track>,
