+++ CommonTools/RecoAlgos/plugins/ConcreteChargedRefCandidateProducer.cc
@@ -11,7 +11,7 @@
  *
  */
 
-#include "CommonTools/RecoAlgos/src/TrackToRefCandidate.h"
+#include "CommonTools/RecoAlgos/interface/TrackToRefCandidate.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
 
 typedef CandidateProducer<reco::TrackCollection, reco::RecoChargedRefCandidateCollection>
