+++ CommonTools/Utils/interface/SelectorPtr.h
