+++ CommonTools/Utils/interface/expressionParser.h
@@ -1,6 +1,6 @@
 #ifndef CommonTools_Utils_expressionParset_h
 #define CommonTools_Utils_expressionParset_h
-#include "CommonTools/Utils/src/ExpressionPtr.h"
+#include "CommonTools/Utils/interface/ExpressionPtr.h"
 #include "CommonTools/Utils/interface/Exception.h"
 #include "FWCore/Reflection/interface/TypeWithDict.h"
 #include <string>
