+++ CommonTools/Utils/src/AnyObjSelector.h
@@ -1,6 +1,6 @@
 #ifndef CommonTools_Utils_AnyObjSelector_h
 #define CommonTools_Utils_AnyObjSelector_h
-#include "CommonTools/Utils/src/SelectorBase.h"
+#include "CommonTools/Utils/interface/SelectorBase.h"
 
 namespace reco {
   namespace parser {
