+++ CommonTools/Utils/src/ExpressionQuaterOperator.h
@@ -10,7 +10,7 @@
  * \version $Revision: 1.2 $
  *
  */
-#include "CommonTools/Utils/src/ExpressionBase.h"
+#include "CommonTools/Utils/interface/ExpressionBase.h"
 #include "CommonTools/Utils/src/ExpressionStack.h"
 
 namespace reco {
