+++ CommonTools/Utils/src/ExpressionVarSetter.cc
@@ -1,6 +1,6 @@
 #include "CommonTools/Utils/src/ExpressionVarSetter.h"
 #include "CommonTools/Utils/src/ExpressionVar.h"
-#include "CommonTools/Utils/src/returnType.h"
+#include "CommonTools/Utils/interface/returnType.h"
 #include "CommonTools/Utils/interface/Exception.h"
 #include <string>
 using namespace reco::parser;
