+++ CommonTools/Utils/src/LogicalUnaryOperator.h
@@ -10,7 +10,7 @@
  * \version $Revision: 1.2 $
  *
  */
-#include "CommonTools/Utils/src/SelectorBase.h"
+#include "CommonTools/Utils/interface/SelectorBase.h"
 #include "CommonTools/Utils/src/SelectorStack.h"
 
 namespace reco {
