+++ CommonTools/Utils/src/MethodInvoker.cc
@@ -3,7 +3,7 @@
 #include "CommonTools/Utils/src/ExpressionVar.h"
 #include "CommonTools/Utils/src/MethodSetter.h"
 #include "CommonTools/Utils/src/findMethod.h"
-#include "CommonTools/Utils/src/returnType.h"
+#include "CommonTools/Utils/interface/returnType.h"
 #include "FWCore/Utilities/interface/EDMException.h"
 
 #include <algorithm>
