+++ CommonTools/Utils/src/TrinarySelector.h
@@ -10,8 +10,8 @@
  * \version $Revision: 1.2 $
  *
  */
-#include "CommonTools/Utils/src/SelectorBase.h"
-#include "CommonTools/Utils/src/ExpressionBase.h"
+#include "CommonTools/Utils/interface/SelectorBase.h"
+#include "CommonTools/Utils/interface/ExpressionBase.h"
 #include "CommonTools/Utils/src/ComparisonBase.h"
 
 namespace reco {
