+++ DQM/Physics/src/QcdUeDQM.cc
@@ -17,7 +17,7 @@
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
 #include "Geometry/CommonDetUnit/interface/GeomDet.h"
-#include "CommonTools/RecoAlgos/src/TrackToRefCandidate.h"
+#include "CommonTools/RecoAlgos/interface/TrackToRefCandidate.h"
 #include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
 #include "Geometry/CommonDetUnit/interface/PixelGeomDetType.h"
 #include "Geometry/CommonDetUnit/interface/PixelGeomDetUnit.h"
