+++ Fireworks/Core/interface/FWExpressionEvaluator.h
@@ -23,10 +23,10 @@
 #include "FWCore/Reflection/interface/TypeWithDict.h"
 
 // user include files
-#include "CommonTools/Utils/src/SelectorPtr.h"
-#include "CommonTools/Utils/src/SelectorBase.h"
-#include "CommonTools/Utils/src/ExpressionPtr.h"
-#include "CommonTools/Utils/src/ExpressionBase.h"
+#include "CommonTools/Utils/interface/SelectorPtr.h"
+#include "CommonTools/Utils/interface/SelectorBase.h"
+#include "CommonTools/Utils/interface/ExpressionPtr.h"
+#include "CommonTools/Utils/interface/ExpressionBase.h"
 
 // forward declarations
