+++ Fireworks/Core/interface/FWItemValueGetter.h
@@ -24,10 +24,10 @@
 #include "Rtypes.h"
 #include "FWCore/Reflection/interface/TypeWithDict.h"
 
-#include "CommonTools/Utils/src/SelectorPtr.h"
-#include "CommonTools/Utils/src/SelectorBase.h"
-#include "CommonTools/Utils/src/ExpressionPtr.h"
-#include "CommonTools/Utils/src/ExpressionBase.h"
+#include "CommonTools/Utils/interface/SelectorPtr.h"
+#include "CommonTools/Utils/interface/SelectorBase.h"
+#include "CommonTools/Utils/interface/ExpressionPtr.h"
+#include "CommonTools/Utils/interface/ExpressionBase.h"
 
 class FWItemValueGetter {
 public:
