+++ PhysicsTools/FWLite/src/classes.h
@@ -1,9 +1,9 @@
 //#include "PhysicsTools/FWLite/interface/FWLiteCandEvaluator.h"
 
-#include "CommonTools/Utils/src/ExpressionPtr.h"
-#include "CommonTools/Utils/src/ExpressionBase.h"
-#include "CommonTools/Utils/src/SelectorPtr.h"
-#include "CommonTools/Utils/src/SelectorBase.h"
+#include "CommonTools/Utils/interface/ExpressionPtr.h"
+#include "CommonTools/Utils/interface/ExpressionBase.h"
+#include "CommonTools/Utils/interface/SelectorPtr.h"
+#include "CommonTools/Utils/interface/SelectorBase.h"
 #include <vector>
 #include "PhysicsTools/FWLite/interface/ScannerHelpers.h"
